Extend explanation of arbitrary formulas for effects
Adding a paragraph because from the introdcution above one can get the
impression that `f: a -> b & ef1, g: b -> c & ef2` just declares that `ef1` is
the effect of `f` and `ef2` is the effect of `g` (which it is in this case and
it is described as such) and then one specifies what the effect of the composed
function is (`-> c & (ef1 and ef2)`). If not explicitly stating that the system
is more general (there must be an assignment of the arbitrary boolean variables
such that the formulas match the purity of the arguments), one can get confused
when reading the more complex example.